v1.7

+++ admin.php
@@ -4,15 +4,22 @@
   if(isset($_POST['btn'])) {
   	if (isset($_FILES)) {
 
-  	  if (!file_exists($dest)) { mkdir($dest); }
+  	  if (! file_exists($dest)) { mkdir($dest); }
       foreach($_FILES as $key => $val) {
         // проверка на имя:
         foreach($val['name'] as $sub_key => $filename) {
+          // проверка существования файла:
+          if (! file_exists($val['tmp_name'][$sub_key])) { 
+            echo "Файл $filename по какой-то причине не загружен на сервер!<br/>";
+            continue; 
+          }
+          // проверка расширения файла:
           preg_match($pattern, $filename, $matches); 
           if (count($matches) < 1) {
           	echo "Файл $filename неверного формата<br/>";
           	continue;
           }
+          // загрузка и проверка загрузки файла в каталог:
           if (move_uploaded_file($val['tmp_name'][$sub_key], $dest.$filename)) {
              echo "Файл $filename успешно загружен в каталог $dest<br/>";
           } else {
@@ -28,25 +35,24 @@
 <html>
 <head>
 	<title>Генератор тестов</title>  
-  <link rel="stylesheet" href="localhost/gentest.css">
+  <link rel="stylesheet" href="./gentest.css">
 	<meta charset="utf-8">
 </head>
 <body>
-  
+  <?php getMainMenu(); ?>
   <form action="" method="POST" enctype="multipart/form-data">
     <fieldset>
     <legend>Загрузчик тестов</legend>
-    <label>Файлы:<br/><input type="file" name="tests[]" multiple required></label><br/><br/>
-    <input type="submit" value="Загрузить тесты" name="btn">
+    <label>Файлы:  <input type="file" name="tests[]" multiple required></label><br/><br/>
+    <input type="submit" value="Загрузить тесты" name="btn"><br/>
     <div class="farea hidden">
       <p>Список выбранных файлов:</p>
-      <ul class="list">
+      <ul class="filelist">
       </ul>
-      <textarea id="fileContent" class="hidden" rows="10" cols="80"></textarea>
+      <textarea id="fileContent" class="hidden" rows="10" cols="180"></textarea>
     </div><br/>
     </fieldset>
   </form>
-  <a href="list.php" target="_blank">Список загруженных тестов</a>
   <script type="text/javascript">
     'use strict';
     let forma=new FormData();
@@ -57,7 +63,7 @@
     function chooseFile(event) {
       const fragment = document.createDocumentFragment();
       const divarea = document.querySelector('.farea');
-      const ularea = document.querySelector('.list');
+      const ularea = document.querySelector('.filelist');
       const files = Array.from(event.target.files);
       Array.from(ularea.children).forEach(lishka => {ularea.removeChild(lishka);}); // удаляем прежние лишки
       divarea.classList.remove('hidden');
@@ -80,6 +86,7 @@
       const reader = new FileReader();
       const fileContent=document.getElementById("fileContent");
       let myfile=forma.get(event.target.textContent);
+      fileContent.classList.remove('hidden');
       reader.addEventListener('load', event=> {
         fileContent.value = event.target.result;
         console.log(fileContent.value);
